Update SQLUndercoverinspectorV1.sql
Update for #93
Datalength increase for column VersionNo in table instanceVersionHistory.

Additional clauses added to not exists to allow inserts in the same say if the version is different to the last logged upgrade during the same day i.e two step upgrades such as upgrade to SP2 then upgrade to CU4 within the same day IF the collections do happen to run during the stepped upgrade.