Update Inspector_BlitzWaits_CustomModule.sql
Commented out the final select statements that show you the current config for the module, these were just annoying when using the powershell installer as they show in the console window. Not updating the updated date on this one as there is no actual change to the module, this is purely a cosmetic change.